drop splash screen activity
