--- samples/ControlCatalog.Android/SplashActivity.cs
@@ -1,35 +1,0 @@
-﻿using Android.App;
-using Android.Content;
-using Android.Content.PM;
-using Android.OS;
-using Avalonia.Android;
-
-namespace ControlCatalog.Android
-{
-    [Activity(Theme = "@style/MyTheme.Splash", MainLauncher = true, NoHistory = true)]
-    public class SplashActivity : AvaloniaSplashActivity<App>
-    {
-        protected override Avalonia.AppBuilder CustomizeAppBuilder(Avalonia.AppBuilder builder)
-        {
-            return base.CustomizeAppBuilder(builder)
-                 .AfterSetup(_ =>
-                 {
-                     Pages.EmbedSample.Implementation = new EmbedSampleAndroid();
-                 });
-        }
-
-        protected override void OnCreate(Bundle? savedInstanceState)
-        {
-            base.OnCreate(savedInstanceState);
-        }
-
-        protected override void OnResume()
-        {
-            base.OnResume();
-
-            StartActivity(new Intent(Application.Context, typeof(MainActivity)));
-
-            Finish();
-        }
-    }
-}
