--- samples/MobileSandbox.Android/SplashActivity.cs
@@ -1,17 +1,0 @@
-﻿using Android.App;
-using Android.Content;
-using Avalonia.Android;
-
-namespace MobileSandbox.Android
-{
-    [Activity(Theme = "@style/MyTheme.Splash", MainLauncher = true, NoHistory = true)]
-    public class SplashActivity : AvaloniaSplashActivity<App>
-    {
-        protected override void OnResume()
-        {
-            base.OnResume();
-
-            StartActivity(new Intent(Application.Context, typeof(MainActivity)));
-        }
-    }
-}
